elk_logs_pipeline/requirements.txt
flask==2.0.1
flask_restful
azure-devops==5.0.0b5
msrest
python-dotenv